web-server/requirements.txt
flask==2.2.2
flask_cors==3.0.10
flask_jwt_extended==4.4.4
flask-restful==0.3.9
requests==2.28.1
sqlalchemy==1.4.42
flask_sqlalchemy==3.0.2
marshmallow==3.18.0
marshmallow_sqlalchemy==0.28.1
pg8000==1.29.3
werkzeug==2.2.2
gunicorn==20.1.0
google-cloud-storage==2.6.0
google-cloud-pubsub==2.13.11
